requirements.txt
flask
boto3
google-generativeai
gradio
python-dotenv